pyproject.toml
[project]
name = "tablefind.ai"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Frank Colson", email = "frankie.colson@gmail.com"},
]
dependencies = [
    "pydantic>=2.8.2",
    "docprompt>=0.8.0",
    "PyYAML>=6.0.2",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "Apache 2.0"}


[tool.pdm]
distribution = false
